refactor(adapters): flatten database/ subdir to match flat adapter convention
The split landed db adapters under adapters/database/, but the documented
convention (and pre-existing agent/contact/chat-summary db adapters) keep
*.database.adapter.ts flat in adapters/. Move the 10 split files up as siblings,
rename _shared.ts -> database.shared.ts, and drop the database/ folder. Barrel +
import paths updated; pure move, no logic change.

Verified: tsc 0 errors, eslint 62 (baseline), adapter naming check passes,
solo adapter specs match the 3 pre-existing baseline failures.
